1.8.1

+++ mainwindow.py
@@ -1657,7 +1657,7 @@ def showhelp():
 
     helpWidget.setLayout(helpLayout)
     helpWindow.setCentralWidget(helpWidget)
-    helpWindow.setFixedSize(helpWindow.frameSize().width() * 0.9, helpWindow.frameSize().height() * 1.5)
+    helpWindow.setFixedSize(int(helpWindow.frameSize().width() * 0.9), int(helpWindow.frameSize().height() * 1.5))
     helpWindow.setWindowTitle("帮助")
     helpWindow.setWindowIcon(QtGui.QIcon(iconPath))
     helpWindow.show()
